Refactoring code #1
