--- src/components/App.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import NavigationBar from './menu/NavigationBar';
-
-class App extends React.Component {
-  render() {
-    return (
-      <NavigationBar />
-    );
-  }
-}
-
-export default App;
