+++ src/components/Login/LoginPage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoginForm from './LoginForm'
+import LoginForm from '../../containers/Login/LoginForm'
 
 class LoginPage extends React.Component {
   render() {
