+++ src/containers/App.js
@@ -1,17 +1,16 @@
 import React from 'react';
 import { Link, Route, Switch } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { logout } from '../../actions/authActions';
+import { logout } from '../actions/authActions';
 
-import LinksPage from '../links/LinksPage';
-import LinkForm from '../shortener/LinkForm';
-import SignupPage from '../signup/SignupPage';
-import LoginPage from '../login/LoginPage';
-import NotFound from '../notfound/NotFound';
-import FlashMessagList from '../flash/FlashMessagesList';
+import LinksPage from './Links/LinksPage';
+import LinkForm from './Shortener/ShortenerForm';
+import SignupPage from '../components/Signup/SignupPage';
+import LoginPage from '../components/Login/LoginPage';
+import NotFound from '../components/Notfound/NotFound';
+import FlashMessagList from './Flash/FlashMessagesList';
 
-
-import requireAuth from '../../utils/requireAuth';
+import requireAuth from '../utils/requireAuth';
 
 const ActiveLink = ({ label, to, activeOnlyWhenExact }) => (
   <Route path={to} exact={activeOnlyWhenExact} children={({ match }) => (
@@ -19,8 +18,7 @@ const ActiveLink = ({ label, to, activeOnlyWhenExact }) => (
   )} />
 );
 
-class NavigationBar extends React.Component {
-
+class App extends React.Component {
   logout = (e) => {
     e.preventDefault();
 
@@ -67,15 +65,15 @@ class NavigationBar extends React.Component {
   }
 }
 
-NavigationBar.propTypes = {
+App.propTypes = {
   auth: React.PropTypes.object.isRequired,
-  logout: React.PropTypes.func.isRequired
+  logout: React.PropTypes.func.isRequired,
 };
 
 function mapStateToProps(state) {
   return {
-   auth: state.auth
+    auth: state.auth,
   };
 }
 
-export default connect(mapStateToProps, { logout })(NavigationBar);
+export default connect(mapStateToProps, { logout })(App);
